refactor: extract form logic to separate hook
